+++ freight/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "freight.apps.FreightConfig"
 
-__version__ = "1.1.4"
+__version__ = "1.2.0"
 __title__ = "Freight"