+++ freight/migrations/0014_auto_20191214_1712.py
@@ -12,5 +12,8 @@ class Migration(migrations.Migration):
     ]
 
     operations = [
-        migrations.RenameModel(old_name="EveOrganization", new_name="EveEntity",),
+        migrations.RenameModel(
+            old_name="EveOrganization",
+            new_name="EveEntity",
+        ),
     ]